requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.4
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
build==0.10.0
    # via pip-tools
cffi==1.15.1
    # via pynacl
charset-normalizer==3.2.0
    # via aiohttp
click==8.1.4
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
discord==2.3.1
    # via -r requirements.in
discord-py==2.3.1
    # via discord
ffmpeg-python==0.2.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
future==0.18.3
    # via ffmpeg-python
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.25.1
    # via -r requirements.in
packaging==23.1
    # via build
pip-tools==6.14.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydub==0.25.1
    # via -r requirements.in
pynacl==1.5.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
pytube==15.0.0
    # via -r requirements.in
wheel==0.40.0
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
